More work on fan configuration
Added fan configuration CAN message
Split M950 message into separate messages for fans, heaters and GPIO

+++ src/CAN/CommandProcessor.cpp
@@ -64,7 +64,7 @@ pre(buf->id.MsgType() == CanMessageType::FirmwareBlockRequest)
 					CanInterface::SendResponse(buf);
 					fileOffset += lengthToSend;
 					lreq -= lengthToSend;
-					if (bytesSent == bytesRead)
+					if ((int32_t)bytesSent == bytesRead)
 					{
 						break;
 					}
